changed int to long for calculating for bigger numbers.

+++ LearningJAVA/Backtracking/GridwaysDirectMethod.java
@@ -1,17 +1,17 @@
 import java.util.Scanner;
 
 public class GridwaysDirectMethod {
-    public static int factorial(int n){
-        int res=1;
+    public static long factorial(int n){
+        long res=1;
         for(int i=1;i<=n;i++){
             res=res*i;
         }
         return res;
     }
-    public static int gridWays(int n, int m){
-        int num=factorial(n-1+m-1);
-        int deno=factorial(n-1)*factorial(m-1);
-        int ways=num/deno;
+    public static long gridWays(int n, int m){
+        long num=factorial(n-1+m-1);
+        long deno=factorial(n-1)*factorial(m-1);
+        long ways=num/deno;
         return ways;
 
     }
@@ -22,7 +22,7 @@ public class GridwaysDirectMethod {
         System.out.print("Enter no. of columns: ");
         int col=sc.nextInt();
         sc.close();
-        int tWays= gridWays(row,col);
+        long tWays= gridWays(row,col);
         System.out.println("Total no. of ways to reach from SRC to DST with only down and right are: "+tWays);
     }
